Split API views into multiple views
